+++ torii/util/tracer.py
@@ -1,29 +1,16 @@
 # SPDX-License-Identifier: BSD-2-Clause
 
-from opcode    import opname
-from sys       import _getframe, implementation
-from typing    import TYPE_CHECKING
+from opcode        import opname
+from sys           import _getframe, implementation
+from typing        import TYPE_CHECKING
 
-from ..        import diagnostics
-from .._typing import SrcLoc
+from ..diagnostics import NameNotFound
+from .._typing     import SrcLoc
 
 __all__ = (
 	'get_src_loc',
 	'get_var_name',
 )
-
-def __dir__() -> list[str]:
-	return list({*__all__, 'NameNotFound'})
-
-def __getattr__(name: str):
-	if name == 'NameNotFound':
-		from warnings import warn
-		warn(
-			f'The import of {name} from {__name__} has been deprecated and moved '
-			f'to torii.diagnostics.{name}', DeprecationWarning, stacklevel = 2
-		)
-		return diagnostics.NameNotFound
-	raise AttributeError(f'Module {__name__!r} has no attribute {name!r}')
 
 _IS_PYPY = implementation.name == 'pypy'
 
@@ -67,7 +54,7 @@ def get_var_name(depth: int = 2, default: str | object = _raise_exception) -> st
 		'CALL_FUNCTION', 'CALL_FUNCTION_KW', 'CALL_FUNCTION_EX', 'CALL_METHOD', 'CALL_METHOD_KW', 'CALL', 'CALL_KW'
 	):
 		if default is _raise_exception:
-			raise diagnostics.NameNotFound
+			raise NameNotFound
 		else:
 			if TYPE_CHECKING:
 				assert isinstance(default, str)
@@ -105,7 +92,7 @@ def get_var_name(depth: int = 2, default: str | object = _raise_exception) -> st
 			index += 2
 		else:
 			if default is _raise_exception:
-				raise diagnostics.NameNotFound
+				raise NameNotFound
 			else:
 				if TYPE_CHECKING:
 					assert isinstance(default, str)